Extract corefile generation from  boshdn
