+++ pkg/kube/util/names/names.go
@@ -40,3 +40,8 @@ func ServiceName(igName string, maxLength int) string {
 	s := names.DNSLabelSafe(igName)
 	return names.TruncateMD5(s, maxLength)
 }
+
+// HeadlessServiceName see interface.
+func HeadlessServiceName(instanceGroupName string) string {
+	return ServiceName(instanceGroupName, 63)
+}
